refactor: rename pros_alloc module to allocator
